refactor(tunnel): isolate pending listener ownership
