+++ native/rust/crates/ripdpi-tunnel-core/src/io_loop/tcp_accept/listener.rs
@@ -1,8 +1,6 @@
 use std::collections::HashMap;
-use std::time::Instant as StdInstant;
 
-use ripdpi_flow_app_attribution::FlowAttributionToken;
-use smoltcp::iface::{SocketHandle, SocketSet};
+use smoltcp::iface::SocketSet;
 use smoltcp::socket::AnySocket;
 use smoltcp::socket::tcp::{self, Socket as TcpSocket};
 use tracing::{debug, warn};
@@ -14,45 +12,13 @@ use super::eviction::evict_oldest_pending_listen;
 use super::socketaddr_to_listen_endpoint;
 
 mod maintenance;
+mod pending;
 
 pub(crate) use maintenance::{gc_stale_pending_listens, reconcile_pending_listeners};
+pub(crate) use pending::PendingListener;
 
 /// Each pending TCP socket owns two `TCP_SOCKET_BUF` allocations. Keep the half-open handshake budget at roughly 16 MiB on Android.
 const MAX_PENDING_LISTENS: usize = 128;
-
-pub(crate) struct PendingListener {
-    pub(crate) handle: SocketHandle,
-    pub(crate) created_at: StdInstant,
-    attribution: PendingAttribution,
-}
-
-/// The listener owns the registration until admission transfers it to a session.
-struct PendingAttribution(Option<FlowAttributionToken>);
-
-impl Drop for PendingAttribution {
-    fn drop(&mut self) {
-        if let Some(token) = self.0.take() {
-            ripdpi_flow_app_attribution::evict_flow(token);
-        }
-    }
-}
-
-impl PendingListener {
-    pub(crate) fn new(handle: SocketHandle, key: TcpFlowKey) -> Self {
-        let token = ripdpi_flow_app_attribution::note_flow(crate::uid_policy::PROTO_TCP, key.src, key.dst).token;
-        Self { handle, created_at: StdInstant::now(), attribution: PendingAttribution(Some(token)) }
-    }
-
-    pub(crate) fn attribution_token(&self) -> &FlowAttributionToken {
-        // Infallible: only admission takes the token, after removing this listener from the pending map.
-        self.attribution.0.as_ref().expect("pending listener owns its attribution")
-    }
-
-    pub(crate) fn take_attribution(&mut self) -> FlowAttributionToken {
-        // Infallible: admission removes the unique listener and calls this exactly once before dropping it.
-        self.attribution.0.take().expect("attribution transfers once at admission")
-    }
-}
 
 pub(crate) fn ensure_pending_listen_for_syn(
     pkt: &[u8],
